Update patch set 2
Patch Set 2: Code-Review-1

Here is some feedback after a discussion with Ricardo on IRC:

I find the final setup a it convoluted or even "backwards": the file will be in /etc/neutron/networking_bgpvpn.conf, and /usr/share/neutron/server/networking_bgpvpn.conf will point to it, and neutron is told to look at a few dirs under /etc/neutron/xxx and also at /usr/share/neutron/server.

I think that usually, what the services read and where configuration is tweaked is all under /etc/ (which may eventually contain symlinks to stuff in /usr/share for stuff that does not move).

So, why not just copy networking_bgpvpn.conf in one of  /etc/neutron/conf.d/common  and  /etc/neutron/conf.d/neutron-server ?

I understand that this is also the way lbass is packaged, and that there may be a very good reason to do so, but since I'm surprised, I'm asking :)

(-1 just to keep track of the question and make sure that nothing unintendly merges, the intent of making the networking_bgpvpn.conf file available in a place where neutron will find it is good)

Patch-set: 2