Update patch set 2
Patch Set 2: -Code-Review

> > Here is some feedback after a discussion with Ricardo on IRC:
 > >
 > > I find the final setup a it convoluted or even "backwards": the
 > > file will be in /etc/neutron/networking_bgpvpn.conf, and
 > > /usr/share/neutron/server/networking_bgpvpn.conf will point to
 > it,
 > > and neutron is told to look at a few dirs under /etc/neutron/xxx
 > > and also at /usr/share/neutron/server.
 > >
 > > I think that usually, what the services read and where
 > > configuration is tweaked is all under /etc/ (which may eventually
 > > contain symlinks to stuff in /usr/share for stuff that does not
 > > move).
 > >
 > > So, why not just copy networking_bgpvpn.conf in one of
 > > /etc/neutron/conf.d/common  and  /etc/neutron/conf.d/neutron-server
 > > ?
 > >
 > > I understand that this is also the way lbass is packaged, and
 > that
 > > there may be a very good reason to do so, but since I'm
 > surprised,
 > > I'm asking :)
 > >
 > > (-1 just to keep track of the question and make sure that nothing
 > > unintendly merges, the intent of making the networking_bgpvpn.conf
 > > file available in a place where neutron will find it is good)
 > 
 > Thomas, the intent of /etc/neutron/conf.d is to contain user
 > controlled config files.

I guess then the question is: what is user controlled ?

On a platform with just the package installed "manually" (not under TripleO control), then this file should be user controlled, in the sense that the user wants to choose whichever driver will be used.

On a platform under automated config, then possibly this file is not considered anymore "under user control".

 > All files shipped with the packages are
 > not going there, so the only place left is /usr/share/...

Ok, perhaps that can be a choice.

But even then, why not put the file directly under /usr/share (without the symlink) ?


 > I agree with the direction of the patch.
 > 
 > If you think it's not clear the way it is, maybe we should document
 > it somewhere.

I won't recommend anything in particular, as soon as it is globally consistent...

(removing my -1 to not block the thing, but interested in the answer nevertheless)

Patch-set: 2
Reviewer: Gerrit User 361 <361@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0