Create python2/python3 package depending on distro
This patch adapts spec file to build python2 only subpackage
when built in el7 systems or python3 only subpackage when built
in fedora or el>7.

It applies the approach shown in [1]

Also ignore unit tests results until [2] is in current-tripleo
repo, this is required to build networking-bgpvpn with
current-tripleo repo.

[1] https://review.rdoproject.org/r/#/c/15247
[2] https://review.openstack.org/#/c/598347/

Change-Id: Id6bd5caf14aa59fe516eaac5c07fd1f52c7e8a20